feat: handleActionButton with new ConversionState handler
Replaced generic boolean flag `isActivelyConverting` with enum type **ConversionState**, containing two cases: `.ready` and `.converting`.

This will allow for more usability within the code for ambiguous scenarios; ie. If we eventually need to handle specific cases like: `.stopping`, `.preparing` or `.paused`.

Furthermore, this makes the code less prone to misunderstandings with human error; ie. Instead of constantly switching is `isActivelyConverting` between `true` and `false`, it's very clear that the app is `.ready` to convert, or the app is currently `.converting`